Namespace the net stuff separately
