+++ backend/src/service.h
@@ -2,7 +2,6 @@
 #define SNIFF_SERVICE
 
 #include "decrypter.h"
-#include "net_card_manager.h"
 #include "proto/service.grpc.pb.h"
 #include "proto/service.pb.h"
 #include "sniffer.h"
@@ -153,7 +152,7 @@ private:
    * @param[in] chan_info Channel information
    * @param[in] chan_proto Protobuf structure to fill out
    */
-  void fill_wifi_chan_info(const wifi_chan_info &chan_info,
+  void fill_wifi_chan_info(const net::wifi_chan_info &chan_info,
                            proto::ChannelInfo *chan_proto);
 
   std::unordered_map<uuid::UUIDv4, std::unique_ptr<Sniffer>> sniffers;
